log is moved into the library. Now the directory structure consists of directives and libraries in separate dir "cryptarsi"
