+++ src/app/app.module.ts
@@ -1,5 +1,4 @@
 import { NgModule } from '@angular/core';
-import { MaterialModule } from '@angular/material';
 import { AppComponent } from './app.component';
 import { AppCryptarsiModule } from './app-cryptarsi/app-cryptarsi.module';
