+++ src/app/cryptarsi/Hash.ts
@@ -1,4 +1,4 @@
-import { log } from '../log';
+import { log } from './log';
 
 export class WordHash {
     static regex = /[^\x00-\x2F\x3A-\x40\x5B-\x60\x7B-\x7E]{2,15}/g; // TODO: the regex must be more correct
